requirements.txt
anyio==3.7.0
APScheduler==3.10.1
asgiref
asttokens==2.0.8
async-timeout==4.0.2
backcall==0.2.0
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
contourpy==1.0.5
cycler==0.11.0
decorator==5.1.1
Django==4.1
django-redis==5.3.0
exceptiongroup==1.1.2
executing==1.1.0
fonttools==4.37.4
google-api-core==2.11.1
google-api-python-client==2.91.0
google-auth==2.21.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.1
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.2
httplib2==0.22.0
idna==3.4
ipython==8.5.0
jedi==0.18.1
kiwisolver==1.4.4
matplotlib==3.6.0
matplotlib-inline==0.1.6
numpy==1.23.4
oauthlib==3.2.2
packaging==21.3
parso==0.8.3
pickleshare==0.7.5
prompt-toolkit==3.0.31
protobuf==4.23.3
psycopg2-binary==2.9.6
pure-eval==0.2.2
pyaes==1.6.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
Pygments==2.13.0
pyparsing==3.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyX==0.15
redis==4.5.5
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scapy==2.4.5
scipy==1.9.3
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.19
sqlparse
stack-data==0.5.1
Telethon==1.28.5
traitlets==5.4.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.0.1
uritemplate==4.1.1
urllib3==1.26.16
uvicorn==0.22.0
wcwidth==0.2.5
